Update delete account function to password

+++ src/Domain/MyProfile/DeleteAccountController.php
@@ -2,18 +2,20 @@
 include_once '../src/Domain/Database.php';
 include '../src/Domain/MyProfile/MyProfileModel.php';
 
-if(isset($_POST['submit'])){
+if (isset($_POST['submit'])) {
     $db = new Database(DATABASE_NAME, DATABASE_USERNAME, DATABASE_PASSWORD);
-    
-    try{
+
+    try {
         $alumni = new MyProfile($db->getConnection(), $_SESSION["alumni"]['alumniId']);
-        $alumni->deleteAccount();
-        session_destroy();
-        header("Location: /login");
-    }catch (Exception $e) {
+        if (password_verify($_POST['deletePassword'], $alumni->getPassword())) {
+            $alumni->deleteAccount();
+            session_destroy();
+            header("Location: /login");
+            exit;
+        }
+    } catch (Exception $e) {
         echo "Exception: " . $e->getMessage();
     }
-    
-}else{
-    header("Location: /myprofile?delete=fail");
 }
+
+header("Location: /myprofile?delete=fail");
